+++ spinn_front_end_common/interface/interface_functions/locate_executable_start_type.py
@@ -1,7 +1,7 @@
 from pacman.model.graphs.machine import MachineVertex
 from spinn_front_end_common.abstract_models.\
     abstract_has_associated_binary import AbstractHasAssociatedBinary
-from spinn_front_end_common.utilities.utility_objs.executable_start_type \
+from spinn_front_end_common.utilities.utility_objs.executable_type \
     import ExecutableType
 from spinn_machine import CoreSubsets
 from spinn_utilities.progress_bar import ProgressBar
