+++ spinn_front_end_common/utilities/utility_objs/__init__.py
@@ -1,5 +1,5 @@
 from .executable_finder import ExecutableFinder
-from .executable_start_type import ExecutableType
+from .executable_type import ExecutableType
 from .live_packet_gather_parameters import LivePacketGatherParameters
 from .provenance_data_item import ProvenanceDataItem
