+++ spinn_front_end_common/utilities/utility_objs/executable_type.py
