+++ spinn_front_end_common/utility_models/chip_power_monitor_machine_vertex.py
@@ -24,7 +24,7 @@ from spinn_front_end_common.interface.buffer_management.buffer_models.\
     AbstractReceiveBuffersToHost
 from spinn_front_end_common.utilities import globals_variables
 from spinn_front_end_common.utilities.utility_objs.\
-    executable_start_type import ExecutableType
+    executable_type import ExecutableType
 from spinn_front_end_common.utilities import constants
 from spinn_front_end_common.utilities import helpful_functions
 from spinn_front_end_common.interface.simulation import simulation_utilities
